- Relevance only need W2V - Chunks of code have their own functions now
